requirements.txt
opencv-python==4.5.5.62
tqdm==4.61.2
pytorch-msssim==1.0.0
timm==0.9.2
tensorboard==2.11.2
tensorboardx==2.6